multipart2

+++ src/main/java/com/capston/matching_app/dto/OutfitSubmissionRequestDTO.java
@@ -1,7 +1,7 @@
 package com.capston.matching_app.dto;
 
-import lombok.Getter;
 import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -9,11 +9,7 @@ import org.springframework.web.multipart.MultipartFile;
 @Setter
 public class OutfitSubmissionRequestDTO {
     @NotNull
-    private Integer matchRequestId;
-
-    @NotNull
-    private Integer userId;
-
-    @NotNull
-    private MultipartFile file;
+    private Long matchRequestId;
+    @NotNull private Long userId;
+    @NotNull private MultipartFile file;
 }
